+++ src/app/(backend)/api/workflows/memory-user-memory/[[...route]]/route.ts
@@ -1,0 +1,3 @@
+import app from '@/server/workflows-hono/memory-user-memory/app';
+
+export const POST = (request: Request) => app.fetch(request);
